Move API server functions from server.go to api.go
Relocate API-related functions to reduce server.go file size and improve code organization.
